+++ CHANGELOG.md
@@ -21,6 +21,8 @@ All notable changes to this project will be documented in this file, formatted v
 - Multiple spaces between some CSS classes inside markup.
 - Twitter URL encoding issue with special characters in text.
 - Email counts not tracking, props @thartl.
+- Showing "Preserve HTTP Counts" setting when Count Source is None.
+- Enabling various settings by default on initial save.
 
 ## [1.2.0] = 2018-05-23
 ### Added