+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file, formatted via [this recommendation](http://keepachangelog.com/).
 
-## [1.3.0] = 2019-04-XX
+## [1.3.0] = 2019-05-13
 ### Added
 - Filter for changing services used by location: `shared_counts_display_services`.
 - Admin bar stats.
@@ -10,8 +10,9 @@ All notable changes to this project will be documented in this file, formatted v
 - Add support for `fastcgi_finish_request` when updating counts.
 - Specific services can be defined in shortcode via `services` attribute (comma separated).
 
+
 ### Changed
-- Pass post_id to `needs_updating` method.
+- Pass post_id to `needs_updating` method, see #74.
 - Removed support for Google+ and StumbleUpon (RIP).
 - Default `letter-spacing` to normal on button labels.
 - Hide Total Counts button if empty and "Hide Empty Counts" setting is enabled.
@@ -23,6 +24,7 @@ All notable changes to this project will be documented in this file, formatted v
 - Email counts not tracking, props @thartl.
 - Showing "Preserve HTTP Counts" setting when Count Source is None.
 - Enabling various settings by default on initial save.
+- When sorting posts by share count in the admin, posts with zero shares are now included, see #76.
 
 ## [1.2.0] = 2018-05-23
 ### Added